==> scripts/00_dunlopgr_DADA2pipeline.ipynb <==
> ### Pipeline to Process Raw Sequences into Phyloseq Object with DADA2 ###
* Prep for Import to QIIME2  (Combine two index files)
* Import to QIIME2
* Demultiplex
* Denoise and Merge
* Prepare OTU Tables and Rep Sequences  *(Note: sample names starting with a digit will break this step)*
* Classify Seqs

*100% Appropriated from the "Atacama Desert Tutorial" for QIIME2*

### Pipeline can handle both 16S rRNA gene and ITS sequences (in theory)####
* Tested on 515f and 806r
* Tested on ITS1

### Commands to Install Dependencies ####
##### || QIIME2 ||
** Note: QIIME2 is still actively in development, and I've noticed frequent new releases. Check for the most up-to-date conda install file <https://docs.qiime2.org/2017.11/install/native/#install-qiime-2-within-a-conda-environment>

* wget https://data.qiime2.org/distro/core/qiime2-2018.2-py35-linux-conda.yml
* conda env create -n qiime2-2018.2 --file qiime2-2018.2-py35-linux-conda.yml
* source activate qiime2-pipeline




##### || Copyrighter rrn Database ||
* The script will automatically install the curated GreenGenes rrn attribute database
* https://github.com/fangly/AmpliCopyrighter

##### || rpy2 (don't use conda version) ||
* pip install rpy2  

##### || phyloseq ||
* conda install -c r r-igraph 
* Rscript -e "source('http://bioconductor.org/biocLite.R');biocLite('phyloseq')" 

##### || R packages ||
* ape   (natively installed in conda environment)


### Citations ###
* Caporaso, J. G., Kuczynski, J., Stombaugh, J., Bittinger, K., Bushman, F. D., Costello, E. K., *et al.* (2010). QIIME allows analysis of high-throughput community sequencing data. Nature methods, 7(5), 335-336.


* McMurdie and Holmes (2013) phyloseq: An R Package for Reproducible Interactive Analysis and Graphics of Microbiome Census Data. PLoS ONE. 8(4):e61217


* Paradis E., Claude J. & Strimmer K. 2004. APE: analyses of phylogenetics and evolution in R language. Bioinformatics 20: 289-290.


* Angly, F. E., Dennis, P. G., Skarshewski, A., Vanwonterghem, I., Hugenholtz, P., & Tyson, G. W. (2014). CopyRighter: a rapid tool for improving the accuracy of microbial community profiles through lineage-specific gene copy number correction. Microbiome, 2(1), 11.

###### Last Modified by R. Wilhelm on October 12th, 2017 ######


# Step 1: User Input

import os, re, numpy as np

# Provide the directory for your index and read files (you can do multiple independently in one go)
growthrate = '/home/cassi/growthrate2018/'
# CW: this is the directory where I want to put my outputs from the pipeline

seqdir = '/home/backup_files/raw_reads/growthrate.cassi.2018/'
# CW: this is the directory where my raw files live

# Prepare an object with the name of the library, the name of the directory object (created above), and the metadatafile name
#datasets = [['name',directory1,'metadata1','domain of life',raw reads directory],['name',directory2,'metadata2','domain of life']]
datasets = [['growthrate',growthrate,'growthrate_metadata.tsv','bacteria',seqdir]]

# Ensure your reads files are named accordingly (or modify to suit your needs)
readFile1 = '88202_C56H8_GR_Library_S1_R1_001.fastq.gz'
readFile2 = '88202_C56H8_GR_Library_S1_R2_001.fastq.gz'
indexFile1 = '88202_C56H8_GR_Library_S1_I1_001.fastq.gz'
indexFile2 = '88202_C56H8_GR_Library_S1_I2_001.fastq.gz'

# Set # of Processors to Use
processors = "20"
# CW: there are 40 processers total, check with others to see how many are being used, affects speed of pipeline

# Classification Database to Use 
# options: "Silva" [default] | "GreenGenes" 
db = "Silva"
# CW: use Silva for bacteria, need UNITE or ITS1db databases for fungi

## Enter Minimum Support for Keeping QIIME Classification
# Note: Classifications that do not meet this criteria will simply be retained, but labeled 'putative'
min_support = 0.8

# Step 2: Concatenate Barcodes for QIIME2 Pipeline

## Note: QIIME takes a single barcode file. The command 'extract_barcodes.py' concatenates the forward and reverse read barcode and attributes it to a single read.

# See http://qiime.org/tutorials/processing_illumina_data.html

# Extracts barcodes from index files, outputs into home directory for project

for dataset in datasets:
    directory = dataset[1]
    rawdir = dataset[4]
    index1 = rawdir+indexFile1
    index2 = rawdir+indexFile2
    
    # Run extract_barcodes to merge the two index files
    !python2 /opt/anaconda2/bin/extract_barcodes.py --input_type barcode_paired_end -f $index1 -r $index2 --bc1_len 8 --bc2_len 8 -o $directory/output

    # QIIME2 import requires a directory containing files names: forward.fastq.gz, reverse.fastq.gz and barcodes.fastq.gz 
    #links directories instead of copying
    !ln -s $rawdir$readFile1 $directory/output/forward.fastq.gz
    !ln -s $rawdir$readFile2 $directory/output/reverse.fastq.gz
    
    # linking raw data directory with output folder in pushpull2018
    
    # Gzip the barcodes files (apparently necessary)
    !pigz -p 5 $directory/output/barcodes.fastq

    # Removed orphaned reads files (not needed)
    !rm $directory/output/reads?.fastq


# Step 3: Import into QIIME2

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    
    os.system(' '.join([
        "qiime tools import",
        "--type EMPPairedEndSequences",
        "--input-path "+directory+"output/",
        "--output-path "+directory+"output/"+name+".qza"
    ]))
    
    # This more direct command is broken by the fact QIIME uses multiple dashes in their arguments (is my theory)
    #!qiime tools import --type EMPPairedEndSequences --input-path $directory/output --output-path $directory/output/$name.qza
     

# Step 4: Demultiplex

########
## Note: The barcode you supply to QIIME is now a concatenation of your forward and reverse barcode.
# Your 'forward' barcode is actually the reverse complement of your reverse barcode and the 'reverse' is your forward barcode. The file 'primers.complete.csv' provides this information corresponding to the Buckley Lab 'primer number'
# This quirk could be corrected in how different sequencing facilities pre-process the output from the sequencer

##
## SLOW STEP (~ 2 - 4 hrs)
##

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    metadata = dataset[2]
    
    os.system(' '.join([
        "qiime demux emp-paired",
        "--m-barcodes-file "+directory+metadata,
        "--m-barcodes-column BarcodeSequence",
        "--i-seqs "+directory+"output/"+name+".qza",
        "--o-per-sample-sequences "+directory+"output/"+name+".demux"
    ]))
    
    # This more direct command is broken by the fact QIIME uses multiple dashes in their arguments (is my theory)
    #!qiime demux emp-paired --m-barcodes-file $directory/$metadata --m-barcodes-category BarcodeSequence --i-seqs $directory/output/$name.qza --o-per-sample-sequences $directory/output/$name.demux
    

print ("demultiplex finished!")

# Step 5: Visualize Quality Scores and Determine Trimming Parameters

## Based on the Graph Produced using the Following Command enter the trim and truncate values. Trim refers to the start of a sequence and truncate the total length (i.e. number of bases to remove from end)

# The example in the Atacam Desert Tutorial trims 13 bp from the start of each read and does not remove any bases from the end of the 150 bp reads:
#  --p-trim-left-f 13 \  
#  --p-trim-left-r 13 \
#  --p-trunc-len-f 150 \
#  --p-trunc-len-r 150

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    
    os.system(' '.join([
        "qiime demux summarize",
        "--i-data "+directory+"/output/"+name+".demux.qza",
        "--o-visualization "+directory+"/output/"+name+".demux.QC.summary.qzv"
    ]))
    
    ## Take the output from this command and drop it into:
    #https://view.qiime2.org

wait_for_user = input("The script will now wait for you to input trimming parameters in the next cell. You will need to take the .qzv files for each library and visualize them at <https://view.qiime2.org>. This is hopefully temporary, while QIIME2 developers improve on q2view.\n\n[ENTER ANYTHING. THIS IS ONLY MEANT TO PAUSE THE PIPELINE]")
print("\nThe script is now proceeding. Stay tuned to make sure trimming works.")

print("qsv done")

# Step 6: Trimming Parameters | USER INPUT REQUIRED

## User Input Required
trim_dict = {}

## Input your trimming parameters into a python dictionary for all libraries
#trim_dict["LibraryName1"] = [trim_forward, truncate_forward, trim_reverse, truncate_reverse]
#trim_dict["LibraryName2"] = [trim_forward, truncate_forward, trim_reverse, truncate_reverse]

## Example
trim_dict["growthrate"] = [5, 240, 5, 200]


print("Done.")

# Step 7: Trim, Denoise and Join (aka 'Merge') Reads Using DADA2

##
## SLOW STEP (~ 6 - 8 hrs, IF multithreading is used)
##

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    
    os.system(' '.join([
        "qiime dada2 denoise-paired",
        "--i-demultiplexed-seqs "+directory+"/output/"+name+".demux.qza",
        "--o-table "+directory+"/output/"+name+".table",
        "--o-representative-sequences "+directory+"/output/"+name+".rep.seqs.final",
        "--p-trim-left-f "+str(trim_dict[name][0]),
        "--p-trim-left-r "+str(trim_dict[name][2]),
        "--p-trunc-len-f "+str(trim_dict[name][1]),
        "--p-trunc-len-r "+str(trim_dict[name][3]),
        "--p-n-threads",
        str(processors)
    ]))
    
    # CW: added str() around processors so that function will work, os.system needs a string passed to it because .join takes strings
    # .join concatenating strings into longer string, ' ' using space as delimiter
    # os.system puts the string in the shell and runs the string as a qiime command
    # qiime runs dada2 on the big string you made, using trim parameters you defined above


print("Finished!")

# Step 8: Create Summary of OTUs

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    metadata = dataset[2]
    
    os.system(' '.join([
        "qiime feature-table summarize",
        "--i-table "+directory+"/output/"+name+".table.qza",
        "--o-visualization "+directory+"/output/"+name+".table.qzv",
        "--m-sample-metadata-file "+directory+metadata
    ]))

    os.system(' '.join([
        "qiime feature-table tabulate-seqs",
        "--i-data "+directory+"/output/"+name+".rep.seqs.final.qza",
        "--o-visualization "+directory+"/output/"+name+".rep.seqs.final.qzv"
    ])) 

# Step 9: Make Phylogenetic Tree

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    metadata = dataset[2]
    domain = dataset[3]

    if domain != "fungi":
        # Generate Alignment with MAFFT
        os.system(' '.join([
            "qiime alignment mafft",
            "--i-sequences "+directory+"/output/"+name+".rep.seqs.final.qza",
            "--o-alignment "+directory+"/output/"+name+".rep.seqs.aligned.qza",
            "--p-n-threads",
            str(processors)
        ]))
        # CW: changed processors into a string again

        # Mask Hypervariable parts of Alignment
        os.system(' '.join([
            "qiime alignment mask",
            "--i-alignment "+directory+"/output/"+name+".rep.seqs.aligned.qza",
            "--o-masked-alignment "+directory+"/output/"+name+".rep.seqs.aligned.masked.qza"
        ])) 

        # Generate Tree with FastTree
        os.system(' '.join([
            "qiime phylogeny fasttree",
            "--i-alignment "+directory+"/output/"+name+".rep.seqs.aligned.masked.qza",
            "--o-tree "+directory+"/output/"+name+".rep.seqs.tree.unrooted.qza",
            "--p-n-threads",
            processors
        ])) 

        # Root Tree
        os.system(' '.join([
            "qiime phylogeny midpoint-root",
            "--i-tree "+directory+"/output/"+name+".rep.seqs.tree.unrooted.qza",
            "--o-rooted-tree "+directory+"/output/"+name+".rep.seqs.tree.final.qza"
        ])) 


# Step 10: Classify Seqs

### Example of creating Silva Classifier DB  (very slow: use a pre-built one if possible)

#### For Silva, remove all the alignment information (unsure of the impact of keeping it) using the following python code:
```python 
import re
from Bio import SeqIO

output = open("silva.nr_v128.fasta", "w")

for record in SeqIO.parse(open("silva.nr_v128.align", "rU"), "fasta") :
    seq = str(record.seq)
    seq = re.sub("\.|-","",seq)  # Remove "." and "-"

    output.write(">"+record.id+"\n"+seq+"\n")```

#### Import fasta sequence file and taxonomy file as .qza
```bash
qiime tools import
  --type 'FeatureData[Sequence]'
  --input-path silva.nr_v128.fasta
  --output-path silva.nr_v128.qza

qiime tools import 
  --type 'FeatureData[Taxonomy]' 
  --source-format HeaderlessTSVTaxonomyFormat 
  --input-path silva.nr_v128.tax 
  --output-path silva.nr_v128.taxonomy.qza```

#### Run QIIME2 'fit-classifier-naive-bayes'
```bash
qiime feature-classifier fit-classifier-naive-bayes 
  --i-reference-reads silva.nr_v128.qza 
  --i-reference-taxonomy silva.nr_v128.taxonomy.qza 
  --o-classifier silva.nr_v128.nb.classifier.qza```

## Note: Different QIIME2 versions can conflict with previously donwloaded databases. This section might have to be updated.
try:
    if db == "GreenGenes":
        classification_db = "/home/db/GreenGenes/qiime2_13.8.99_515.806_nb.classifier.qza"
    else:
        classification_db = "/home/db/Silva/silva.nr_v128.nb.classifier.qza"
        
except:
        classification_db = "/home/db/Silva/silva.nr_v128.nb.classifier.qza"
        
for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    metadata = dataset[2]
    domain = dataset[3]

    # Classify
    if domain == 'bacteria':
        os.system(' '.join([
            "qiime feature-classifier classify-sklearn",
            "--i-classifier",
            classification_db,
            "--i-reads "+directory+"/output/"+name+".rep.seqs.final.qza",
            "--o-classification "+directory+"/output/"+name+".taxonomy.final.qza",
            "--p-n-jobs",
            processors
        ]))

    if domain == 'fungi':
        os.system(' '.join([
            "qiime feature-classifier classify-sklearn",
            "--i-classifier /home/db/UNITE/qiime2_unite_ver7.99_20.11.2016_classifier.qza",
            "--i-reads "+directory+"/output/"+name+".rep.seqs.final.qza",
            "--o-classification "+directory+"/output/"+name+".taxonomy.final.qza",
            "--p-n-jobs",
            processors
        ]))

    # Output Summary
    os.system(' '.join([
        "qiime metadata tabulate",
        "--m-input-file "+directory+"/output/"+name+".taxonomy.final.qza",
        "--o-visualization "+directory+"/output/"+name+".taxonomy.final.summary.qzv"
    ])) 

# Step 11: Prepare Data for Import to Phyloseq

## Make Function to Re-Format Taxonomy File to Contain Full Column Information 
# and factor in the certain of the taxonomic assignment

def format_taxonomy(tax_file, classification_db, min_support):
    output = open(re.sub(".tsv",".fixed.tsv",tax_file), "w")

 
    # Silva db lacks species classifications
    if classification_db == "GreenGenes":
        full_rank_length = 7
        output.write("\t".join(["OTU","Domain","Phylum","Class","Order","Family","Genus","Species"])+"\n")
    else:
        full_rank_length = 6  
        output.write("\t".join(["OTU","Domain","Phylum","Class","Order","Family","Genus"])+"\n")
        
    with open(tax_file, "r") as f:
        next(f) #skip header

        for line in f:
            line = line.strip()
            line = line.split("\t")

            read_id = line[0]
            tax_string = line[1]

            ## Remove All Underscore Garbage (I need aesthetics)
            if classification_db == "GreenGenes":
                tax_string = re.sub("k__|p__|c__|o__|f__|g__|s__","",tax_string)
            else:
                tax_string = re.sub("_cl|_or|_fa|_ge","",tax_string)
            
            # Split full rank into ranks
            full_rank = tax_string.split(";")
          
            # Getting trailing empty tab in Silva
            if full_rank[len(full_rank)-1] == "":
                    full_rank = full_rank[:-1]
                    
            ## Identify the Lowest Classified Taxonomic Rank
            # Account for cases when a taxonomic rank contains an empty space (common in GreenGenes output)
            last_classified = full_rank[len(full_rank)-1]            

            count = 1
            while last_classified == " ":
                last_classified = full_rank[len(full_rank)-count]
                count = count + 1

            # Annotate the last classified as 'putative' if it does not meet the minimum support criteria
            # Older versions of this script contain code to designate all taxonomic ranks as 'putative' in this case, but 
            # this seems conservative
            if float(line[2]) < float(min_support):
                    full_rank[full_rank.index(last_classified)] = "putative "+last_classified
                    last_classified = "putative "+last_classified
                    
            # Add in columns containing unclassified taxonomic information
            try: # In Silva, many classifications are a single entry (which breaks from the reliance on lists for full_rank.index)
                for n in range(full_rank.index(last_classified)+1, full_rank_length, 1):               
                    try:
                        full_rank[n] = "unclassified "+last_classified
                    except:
                        full_rank.append("unclassified "+last_classified)
            except:
                for n in range(0, full_rank_length, 1):               
                    try:
                        full_rank[n] = "unclassified "+last_classified
                    except:
                        full_rank.append("unclassified "+last_classified)
                    
            # Clean-up the trailing whitespace introduced in Silva classification 
            if classification_db == "Silva":
                full_rank = [x.strip(' ') for x in full_rank]

            # Write Taxonomy to File
            output.write(read_id+"\t"+'\t'.join(full_rank)+"\n")
            
    return()

#####################
## Export from QIIME2

# CW: had to remove first / in front of every output

for dataset in datasets:
    name = dataset[0]
    directory = dataset[1]
    metadata = dataset[2]
    domain = dataset[3]

    ## Final Output Names
    fasta_file = directory+"/output/"+name+".rep.seqs.final.fasta"
    tree_file = directory+"/output/"+name+".tree.final.nwk"
    tax_file = directory+"/output/"+name+".taxonomy.final.tsv"
    count_table = directory+"/output/"+name+".counts.final.biom"

    # Export Classifications
    os.system(' '.join([
        "qiime tools export",
        directory+"/output/"+name+".taxonomy.final.qza",
        "--output-dir "+directory+"/output/"
    ]))
    
    # Reformat Classifications to meet phyloseq format   
    format_taxonomy(directory+"/output/taxonomy.tsv", db, min_support)

    # Export SV Table
    os.system(' '.join([
        "qiime tools export",
        directory+"/output/"+name+".table.qza",
        "--output-dir "+directory+"/output/"
    ]))

    # Export SV Sequences
    os.system(' '.join([
        "qiime tools export",
        directory+"/output/"+name+".rep.seqs.final.qza",
        "--output-dir "+directory+"/output/"
    ]))
    
    # Export Tree
    os.system(' '.join([
        "qiime tools export",
        directory+"/output/"+name+".rep.seqs.tree.final.qza",
        "--output-dir "+directory+"/output/"
    ]))
    
    # Rename Exported Files
    %mv $directory/output/dna-sequences.fasta $fasta_file
    %mv $directory/output/feature-table.biom $count_table
    %mv $directory/output/taxonomy.fixed.tsv $tax_file
    
    #if domain == "bacteria":
        #%mv $directory/output/tree.nwk $tree_file
        
    # CW: above broke the code somehow, moved to new cell and worked

%mv $directory/output/tree.nwk $tree_file
