==> scripts/0x_dunlop_growthvsdeath.Rmd <==
---
title: "Untitled"
author: "Cassandra Wattenburger"
date: "12/3/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(results = "show")
knitr::opts_chunk$set(message = FALSE)
```

```{r}
rm(list=ls())

library("tidyverse")

sessionInfo()
```

# Import data

```{r}
# Growth and death estimates
growth <- readRDS("../rdata.files/gr_gr.paprica.clean.rds") %>% 
  rename(gr_start_day=start_day, gr_end_day=end_day,
         gr_start_abund=start_abund, gr_end_abund=end_abund) %>% 
  mutate(Label = paste0(Soil, Amendment, Replicate, ASV),
         gr_change_abund = gr_end_abund - gr_start_abund,
         gr_length = gr_end_day - gr_start_day)

death <- readRDS("../rdata.files/gr_death_estimates.rds") %>% 
  select(everything(), -c(start_pt, end_pt, Domain:Genus, decay)) %>% 
  rename(de_rate=death_rate, de_start_day=start_day, de_end_day=end_day, de_length = length_days,
         de_start_abund=start_abund, de_end_abund=end_abund, de_change_abund = change_abund)

# Normalized abundances
norm <- readRDS("../rdata.files/gr_ucosm.norm.clean.rds") %>% 
  mutate(Label = paste0(Soil, Amendment, Replicate, ASV)) %>% 
  select(Label, Soil, Amendment, Replicate, ASV, Day, Domain:Genus, norm_abund) %>% 
  filter(norm_abund > 0)

# Taxa with estimated growth and then death
growth_death <- inner_join(growth, death) %>% 
  select(Label, Soil, Amendment, Replicate, Domain:Genus, ASV,
         k, g, gr_start_day, gr_end_day, gr_start_abund, gr_end_abund, gr_change_abund, gr_length,
         de_rate, half_time, de_start_day, de_end_day, de_start_abund, de_end_abund, de_change_abund, de_length,
         edge_num:n16S)
```

# Descriptive stats

### How many death estimates?

Total:

```{r}
nrow(growth_death)
```

```{r}
length(as.character(unique(growth_death$ASV)))
```

### Number of ASVs per treatment

```{r}
death_asvs <- growth_death %>%
  group_by(Soil, Amendment) %>% 
  summarize(num_est = length(unique(ASV))) %>%
  ungroup()
  
death_asvs
```

### Overall descriptive stats

Start, end, and length of death:

```{r}
death_overall <- growth_death %>%
  group_by() %>% 
  summarize(de_start_day_min = min(de_start_day),
            de_start_day_max = max(de_start_day),
            de_start_day_mean = mean(de_start_day),
            de_start_day_sd = sd(de_start_day),
            de_end_day_min = min(de_end_day),
            de_end_day_max = max(de_end_day),
            de_end_day_mean = mean(de_end_day),
            de_end_day_sd = sd(de_end_day)) %>%
  ungroup()
  
death_overall
```

Halving times:

```{r}
death_g <- growth_death %>% 
  group_by() %>% 
  summarize(half_time_min = min(half_time),
            half_time_max = max(half_time),
            half_time_mean = mean(half_time),
            half_time_sd = sd(half_time)) %>% 
  ungroup()

death_g
```

### Phyla represented

```{r}
# Number of phyla
length(unique(as.character(growth_death$Phylum)))

# Which phyla
death_phyla <- growth_death %>% 
  group_by(Phylum) %>% 
  summarize(num_est = n()) %>%
  ungroup() %>% 
  arrange(desc(num_est))
  
death_phyla
```

# Aggregate parameters

Overview of soil treatment affects on measured community overall.

### Proportion of taxa that grew then died vs just grew

```{r}
count_growth <- growth %>% 
  group_by(Soil, Amendment, Replicate) %>% 
  summarize(growth_count = n()) %>% 
  ungroup()

count_death <- death %>% 
  group_by(Soil, Amendment, Replicate) %>% 
  summarize(death_count = n()) %>% 
  ungroup()

# Proportion
prop_death <- inner_join(count_growth, count_death) %>% 
  mutate(prop =death_count/growth_count)

# Graph
prop_death %>% 
  ggplot(aes(x=Soil, y=prop, color=Amendment)) +
  geom_point() +
  theme_test()
```

Highly variable.

```{r}
# Aggregate parameters
growth_death_agg <- growth_death %>% 
  group_by(Soil, Amendment, Replicate) %>% 
    summarize(k = mean(k), g = mean(g),
            gr_start_day = mean(gr_start_day), gr_end_day = mean(gr_end_day), 
            gr_length = mean(gr_length),
            gr_start_abund_corr = mean(gr_start_abund/n16S, na.rm = TRUE), # 16S copy num correction
            gr_change_abund_corr = mean(gr_change_abund/n16S, na.rm = TRUE), # 16S copy num correction
            de_rate = mean(de_rate),
            de_start_day = mean(de_start_day), de_end_day = mean(de_end_day), 
            de_length = mean(de_length),
            de_start_abund_corr = mean(de_start_abund/n16S, na.rm = TRUE), # 16S copy num correction
            de_change_abund_corr = mean(de_change_abund/n16S, na.rm = TRUE)) %>% # 16S copy num correction
  ungroup() 
```

### Death rate

```{r}
growth_death_agg %>%
  ggplot(aes(x=Soil, y=de_rate, color=Amendment)) +
  geom_point() +
  theme_test()
```

Highly variable, not seeing a trend.

### Death change abundance

```{r}
growth_death_agg %>% 
  ggplot(aes(x=Soil, y=abs(de_change_abund_corr), color=Amendment)) +
  geom_point() +
  theme_test()
```

Seems like trend of greater changes in abundance due to death in amended soils?

**Statistics**

Linear model:

```{r}
chabund_agg_lm <- lm(abs(de_change_abund_corr) ~ Soil + Amendment + Soil*Amendment, data=growth_death_agg)
hist(resid(chabund_agg_lm)) # normal enough
plot(predict(chabund_agg_lm), resid(chabund_agg_lm)) # unequal variance, log transform makes worse
```

I don't like the unequal variance here, even with a log transform. I'll use a Welch's t-test to test amendment.

Welch's t-test:

* Non-transformed data

```{r}
# Cropped
chabund_agg_welch <- t.test(abs(de_change_abund_corr) ~ Amendment, data=growth_death_agg, var.equal=FALSE)
chabund_agg_welch
```

# ASV level

Average each ASV across replicates to get a parameter estimate for each taxa in each treatment. This allows a more nuanced investigation of how soil treatments interact with growth responses of individual ASVs/populations rather than the community as a whole.

```{r}
# Average growth estimates for each ASV
growth_asv <- growth %>%
  group_by(ASV, Soil, Amendment) %>% 
  summarize(k = mean(k), g = mean(g),
            gr_start_day = mean(gr_start_day), gr_end_day = mean(gr_end_day), 
            gr_length = mean(gr_length),
            gr_start_abund_corr = mean(gr_start_abund/n16S, na.rm=TRUE),
            gr_change_abund_corr = mean(gr_change_abund/n16S, na.rm=TRUE),
            n16S = mean(n16S, na.rm=TRUE)) %>% 
  ungroup()

# Average death estimates for each ASV
# Logarithmic only
death_asv_log <- death %>% 
  group_by(ASV, Soil, Amendment) %>% 
  summarize(de_rate = mean(de_rate),
            de_start_day = mean(de_start_day), de_end_day = mean(de_end_day), 
            de_length = mean(de_length),
            de_start_abund = mean(de_start_abund), de_end_abund = mean(de_end_abund), 
            de_change_abund = mean(de_change_abund)) %>% 
  ungroup()

# Both growth and death
growth_death_asv <- growth_death %>%
  group_by(ASV, Soil, Amendment) %>% 
  summarize(k = mean(k), g = mean(g),
            gr_start_day = mean(gr_start_day), gr_end_day = mean(gr_end_day), 
            gr_length = mean(gr_length),
            gr_start_abund_corr = mean(gr_start_abund/n16S, na.rm=TRUE),
            gr_change_abund_corr = mean(gr_change_abund/n16S, na.rm=TRUE),
            de_rate = mean(de_rate),
            de_start_day = mean(de_start_day), de_end_day = mean(de_end_day), 
            de_length = mean(de_length),
            de_start_abund_corr = mean(de_start_abund/n16S, na.rm=TRUE),
            de_change_abund_corr = mean(de_change_abund/n16S, na.rm=TRUE),
            n16S = mean(n16S, na.rm=TRUE)) %>% 
  ungroup()
```

### Growth rate vs did death happen at all

Hypothesis: Taxa with faster growth rates more likely to experience death afterwards
* boom/bust lifestyle
* didn't sample enough/long enough to capture slower grower death

```{r}
# Presence/absence of death in growing ASVs
no_death_asv <- anti_join(growth_asv, death_asv_log, by=c("ASV", "Soil", "Amendment")) %>% 
  add_column(died ="no")

growth_death_asv <- growth_death_asv %>% 
  add_column(died="yes")

# Include ASVs without detected death after growth
all_asv <- bind_rows(no_death_asv, growth_death_asv) %>% 
  mutate()
```

```{r}
# Graph
all_asv %>% 
  ggplot(aes(x=died, y=log(k))) +
  geom_boxplot() +
  geom_jitter(alpha=0.5) +
  facet_wrap(Soil~Amendment) +
  theme_test()
```

Greater variability and sample size of taxa that did not have estimated death after growth. Dificult to say whether this difference in variation/growth rate is real or due to undersampling.

### Growth change abundance vs death yes/no

Hypothesis: Taxa that grew more biomass more likely to experience death afterwards
* boom/bust lifestyle
* didn't sample enough/long enough to slower/less growth and death

```{r}
# Graph
all_asv %>% 
  ggplot(aes(x=died, y=log(gr_change_abund_corr))) +
  geom_boxplot() +
  geom_jitter(alpha=0.5) +
  facet_wrap(Soil~Amendment) +
  theme_test()
```

### Growth rate vs death rate

Hypothesis: Taxa with faster growth rate more likely to experience faster death
* boom/bust lifestyle

```{r}
# Graph
growth_death_asv %>% 
  ggplot(aes(x=log(k), y=log(abs(de_rate)))) +
  geom_smooth(method="lm", linetype=2, alpha=0.5) +
  geom_point(alpha=0.5) +
  facet_wrap(Soil~Amendment) +
  theme_test()
```

Perhaps a trend of slight positive relationship.

**Statistics**

Pearson correlations with multiple test correction

```{r}
# Treatments
growthrate_deathrate_pcor_trts <- data.frame()
for (s in c("C3", "S17")) {
  for (a in c("Y", "N")) {
    data_sub <- filter(growth_death_asv, Soil==s & Amendment==a)
    pcor <- cor.test(data_sub$k, abs(data_sub$de_rate), method="pearson")
    this_row <-  data.frame("Soil"=s, "Amendment"=a, "estimate"=pcor$estimate, "pvalue"=pcor$p.value)
    growthrate_deathrate_pcor_trts <- bind_rows(growthrate_deathrate_pcor_trts, this_row)
  }
}

# Multiple test correction
padj <- p.adjust(growthrate_deathrate_pcor_trts$pvalue, method="holm", n=4)
growthrate_deathrate_pcor_trts <- bind_cols(growthrate_deathrate_pcor_trts, padj=padj)

# Results
growthrate_deathrate_pcor_trts
```

No significance.

### Growth rate vs death change in abundance

Logarithmic only

```{r}
# Graph
growth_death_asv %>% 
  ggplot(aes(x=log(k), y=log(abs(de_change_abund_corr)))) +
  geom_smooth(method="lm", linetype=2, alpha=0.5) +
  geom_point(alpha=0.5) +
  facet_wrap(Soil~Amendment) +
  theme_test()
```

### Growth change in abundance vs death rate

Hypothesis: Taxa that grew more more likely to experience faster death
* boom/bust lifestyle

```{r}
# Graph
growth_death_asv %>% 
  ggplot(aes(x=log(gr_change_abund_corr), y=log(abs(de_rate)))) + # cannot take ln of negative number
  geom_point(alpha=0.5) +
  geom_smooth(method="lm", linetype=2, alpha=0.5) +
  facet_wrap(Soil~Amendment) +
  theme_test()
```

Possibly slight positive correlation in amended soils but negative in unamended? This would mean that in unamended soils, as the change in growth biomass increased, the death rate tended to be lower, but in the amended soils, the opposite was true.

**Statistics**

Pearson correlations with multiple test correction

```{r}
# Treatments
grchabund_deathrate_pcor_trts <- data.frame()
for (s in c("C3", "S17")) {
  for (a in c("Y", "N")) {
    data_sub <- filter(growth_death_asv, Soil==s & Amendment==a)
    pcor <- cor.test(log(data_sub$gr_change_abund_corr), log(abs(data_sub$de_rate)), method="pearson")
    this_row <-  data.frame("Soil"=s, "Amendment"=a, "estimate"=pcor$estimate, "pvalue"=pcor$p.value)
    grchabund_deathrate_pcor_trts <- bind_rows(grchabund_deathrate_pcor_trts, this_row)
  }
}

# Multiple test correction
padj <- p.adjust(grchabund_deathrate_pcor_trts$pvalue, method="holm", n=4)
grchabund_deathrate_pcor_trts <- bind_cols(grchabund_deathrate_pcor_trts, padj=padj)

# Results
grchabund_deathrate_pcor_trts
```

No significance.

### Growth change in abundance vs death change in abundance

Hypothesis: Taxa that accumulated more biomass during growth can and will lose more biomass to death
* boom/bust lifestyle

```{r}
# Graph
growth_death_asv %>%
  filter(died=="yes") %>%
  ggplot(aes(x=log(gr_change_abund_corr), y=log(abs(de_change_abund_corr)))) +
  geom_point(alpha=0.5) +
  geom_smooth(method="lm", linetype=2, alpha=0.5) +
  facet_wrap(Soil~Amendment) +
  theme_test()
```

Positive correlation between change in abundance due to death and growth? If a taxa grew more, it died more. There also seems to be a treatment effect but this could be due to undersampling things that grew then died in the unammended soil.

**Statistics**

Pearson correlations with multiple test correction

```{r}
# Treatments
chabund_growth_death_pcor <- data.frame()
for (s in c("C3", "S17")) {
  for (a in c("Y", "N")) {
    data_sub <- filter(growth_death_asv, Soil==s & Amendment==a)
    pcor <- cor.test(log(data_sub$gr_change_abund_corr), log(abs(data_sub$de_change_abund_corr)), method="pearson")
    this_row <-  data.frame("Soil"=s, "Amendment"=a, "estimate"=pcor$estimate, "pvalue"=pcor$p.value)
    chabund_growth_death_pcor <- bind_rows(chabund_growth_death_pcor, this_row)
  }
}

# Multiple test correction
padj <- p.adjust(chabund_growth_death_pcor$pvalue, method="holm", n=4)
chabund_growth_death_pcor_trts <- bind_cols(chabund_growth_death_pcor, padj=padj)

# Results
chabund_growth_death_pcor
```

Positive correlation between growth and death change in abundance in amended but not unamended soil. More growth means more potential for losses to death?

# Figures

Aggregate death change in abundance

```{r}
plot <- growth_death_agg %>% 
   mutate(Soil = fct_recode(Soil, Cropped = "C3", Successional = "S17"),
          Amendment = as_factor(Amendment),
         Amendment = fct_recode(Amendment, Resources = "Y", Water = "N"),
         Amendment = fct_relevel(Amendment, c("Resources", "Water"))) %>% 
  ggplot(aes(x=Soil, y=abs(de_change_abund_corr))) +
  geom_jitter(aes(shape=Amendment), width=0.1) +
  scale_shape_manual(values=c(19,1)) +
  theme_classic() +
  theme(axis.title = element_blank(),
        text = element_text(size=9),
        strip.text.x = element_text(size = 10, hjust = 0),
        legend.position = "none",
        strip.background = element_blank())
plot
```

```{r, eval=FALSE}
ggsave(plot, filename="../figures/fig_sample_deathchabund.svg", units="mm", width=85, height=85, device="svg")
```

Growth vs death change in abundance

```{r}
plot <- growth_death_asv %>%
  mutate(Treatment = paste0(Soil, Amendment),
         Treatment = fct_recode(Treatment, 'Cropped, water' = "C3N", 'Cropped, resources' = "C3Y",
                                'Successional, water' = "S17N", 'Successional, resources' = "S17Y"),
         Treatment = factor(Treatment, levels=c("Cropped, water", "Cropped, resources",
                                                 "Successional, water", "Successional, resources"))) %>%
  mutate(decay=="logarithmic") %>% 
  filter(died=="yes") %>%
  ggplot(aes(x=log(gr_change_abund_corr), y=log(abs(de_change_abund_corr)))) +
  geom_point(shape=1, alpha= 0.5) +
  geom_smooth(method="lm", linetype=2, color="black", alpha=0.5) +
  facet_wrap(~Treatment) +
  theme_test() +
  theme(text = element_text(size=9),
        strip.text.x = element_text(size = 9, hjust = 0),
        strip.background = element_blank(),
        axis.title = element_blank())
plot
```

```{r, eval=FALSE}
ggsave(plot, filename = "../figures/fig_growthxdeathchabund.svg", units="mm", width=85, height=85, device="svg")
```
