==> scripts/06_dunlopgr_treatments.Rmd <==
---
title: "Dunlop growth rates - Growth metrics by treatment"
author: "Cassandra Wattenburger"
date: "2/13/2021"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(results = "show")
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```


**Goal of this script:** Investigate the effects of soil habitat on in-situ growth dynamics.


# Import libraries

```{r}
library(tidyverse)
library(cowplot)
library(data.table)
library(phyloseq)
library(lmerTest)

sessionInfo()

rm(list=ls())
```


# Import growth data

```{r}
growth <- readRDS("../rdata.files/gr_gr.paprica.clean.rds")
```

Aggregate growth estimate parameters:

Average across ASVs to get a value for each replicate

```{r}
# Replicate averages
growth.trt <- growth %>%
  group_by(Soil, Amendment, Replicate) %>%
  summarize(start_day = mean(start_day),
            end_day = mean(end_day),
            start_abund_corr = mean(start_abund/n16S, na.rm=TRUE),
            change_abund_corr = mean((end_abund-start_abund)/n16S, na.rm=TRUE),
            k = mean(k),
            g = mean(g),
            n16S = mean(n16S, na.rm=TRUE)) %>% # archaea have no 16S estimates in this version of PAPRICA
  ungroup()
```

Overall:

```{r}
growth_overall <- growth %>%
  summarize(min_start_day = min(start_day),
            max_start_day = max(start_day),
            mean_start_day = mean(start_day),
            sd_start_day = sd(start_day),
            min_end_day = min(end_day),
            max_end_day = max(end_day),
            mean_end_day = mean(end_day),
            sd_end_day = sd(end_day),
            min_g = min(g),
            max_g = max(g),
            mean_g = mean(g),
            sd_g = sd(g)) %>% 
  ungroup()

growth_overall
```

# Start of growth window

The day of the start of the growth window is being used as a proxy for lag time. The later a growth window starts, the longer lag the bacteria experienced before starting growth.

Hypotheses:

1. Bacteria in cropped soils will start growth sooner than those in successional, due to more frequent disturbances (tillage, low plant cover) and/or higher nutrient availability (fertilizers)
1. Bacteria in C-amended soils will start growth sooner than those in the water control due to having more resources available to ramp up growth

```{r}
growth.trt %>%
  ggplot(aes(x=Soil, y=start_day, color=Amendment)) +
  geom_point() +
  labs(title="Start of growth", y="Day") +
  theme_test()
```

**Statistics**

ANOVA

```{r}
strtday.trt.lm = lm(start_day ~ Soil + Amendment + Soil*Amendment, data=growth.trt)
hist(resid(strtday.trt.lm)) # check normality
plot(predict(strtday.trt.lm), resid(strtday.trt.lm)) # check heteroskedasticity
anova(strtday.trt.lm)
```

Post hoc:

* Cropped water vs C-amended
* Successional water vs C-amended
* Cropped vs successional, water
* Cropped vs successional, C-amended

Fit simple linear model and check residuals:

```{r}
# Contrasts
strtday.trt.lm1 <- lm(start_day ~ Amendment, data=growth.trt[growth.trt$Soil=="C3",], var.equal = TRUE)
hist(resid(strtday.trt.lm1)) # check normality
plot(predict(strtday.trt.lm1), resid(strtday.trt.lm1)) # check heteroskedasticity
```

In this contrast, residuals appear normal, but data has heteroskedasticity. Log transform corrects heteroskedasticity, but makes residuals non-normal. I'll conduct Welch's t-tests on the untransformed data.

```{r}
# Welch's t-tests
strtday.trt.welch1 <- t.test(start_day ~ Amendment, data=growth.trt[growth.trt$Soil=="C3",], var.equal = FALSE)
strtday.trt.welch2 <- t.test(start_day ~ Amendment, data=growth.trt[growth.trt$Soil=="S17",], var.equal = FALSE)
strtday.trt.welch3 <- t.test(start_day ~ Soil, data=growth.trt[growth.trt$Amendment=="Y",], var.equal = FALSE)
strtday.trt.welch4 <- t.test(start_day ~ Soil, data=growth.trt[growth.trt$Amendment=="N",], var.equal = FALSE)

# Adjust p-values
strtday.trt.adjpvals <- p.adjust(p <- c(strtday.trt.welch1$p.value, strtday.trt.welch2$p.value, 
                                        strtday.trt.welch3$p.value, strtday.trt.welch4$p.value),
                                        method="BH", n=length(p))

# Results
strtday.trt.results <- data.frame(contrast = c("Cropped, water vs C-amended", 
                                               "Successional, water vs C-amended", 
                                               "C-amended, cropped vs successional", 
                                               "Water, cropped vs successional"), pvals = strtday.trt.adjpvals)
strtday.trt.results
```

Not enough power.

# End of growth window

Hypotheses:

1. Bacteria will end growth sooner in the water control compared to C-amended because less resources are available to sustain growth
1. Toss up on effect of soil habitat, cropped soil disturbances could promote growth that ends sooner due to heightened competition, or longer because greater N and other nutrient availability
1. Null: end growth follows start growth patterns because delayed growth results in later end of growth

```{r}
growth.trt %>%
  ggplot(aes(x=Soil, y=end_day, color=Amendment)) +
  geom_point(alpha=0.5) +
  labs(title="End of growth", y="Day") +
  theme_test()
```

**Statistics**

ANOVA

```{r}
endday.trt.lm = lm(end_day ~ Soil + Amendment + Soil*Amendment, data=growth.trt)
hist(resid(endday.trt.lm)) # check normality
plot(predict(endday.trt.lm), resid(endday.trt.lm))
anova(endday.trt.lm)
```

End of growth is not affected by soil or amendment. Trends are similar to start of growth data, though.

# Growth rate

Hypotheses:

1. Bacteria in cropped soils will have faster growth rates, because of more frequent disturbances that lead to boom-bust lifestyle and/or increased nutrient availability from fertilization
1. Bacteria in C-amended soils will have greater growth rates, because more fuel for growth available

```{r}
growth.trt %>%
  ggplot(aes(x=Soil, y=k, color=Amendment)) +
  geom_point(alpha=0.5) +
  labs(title="Specific growth rate", y="k") +
  theme_test()
```

**Statistics**

ANOVA

```{r}
k.trt.lm = lm(k ~ Soil + Amendment + Soil*Amendment, data=growth.trt)
hist(resid(k.trt.lm)) # check normality
plot(predict(k.trt.lm), resid(k.trt.lm))
anova(k.trt.lm)
```

Trends almost support second hypothesis, but data is too variable.

# Change in abundance

* 16S copy number corrected

Competitive bacteria are able to increase there abundance more when they encounter substrate compared to competing bacteria. There are likely various strategies to be competitive. Some bacteria are likely more competitive and better able to grow under certain environmental or resource conditions but not others.

Hypotheses:

1. Cropped soils will have greater change in abundance due to boom-bust/more nutrients available to lift limitations on growth
1. Bacteria in C-amended soils will have greater change in abundance due to greater resource availability

```{r}
growth.trt %>%
  ggplot(aes(x=Soil, y=change_abund_corr, color=Amendment)) +
  geom_point(alpha=0.5) +
  labs(title="Change in abundance", y="change in abundance") +
  theme_test()
```

**Statistics**

ANOVA

* log transformation improves heteroskedasticity

```{r}
changeabund.trt.lm = lm(change_abund_corr ~ Soil + Amendment + Soil:Amendment, data=growth.trt)
hist(resid(changeabund.trt.lm)) # check normality
plot(predict(changeabund.trt.lm), resid(changeabund.trt.lm)) # check heteroskedasticity
anova(changeabund.trt.lm)
```

Data looks slightly non-normal, but ANOVA is typically robust to moderate deviation.

Post hoc:

* Cropped water vs C-amended
* Successional water vs C-amended

Try simple linear model and check residuals

```{r}
# Cropped water vs C-amended
test1 <- lm(change_abund_corr ~ Amendment, data=growth.trt[growth.trt$Soil=="C3",])
hist(resid(test1))
plot(predict(test1), resid(test1))

# log transform
test2 <- lm(log(change_abund_corr) ~ Amendment, data=growth.trt[growth.trt$Soil=="C3",])
hist(resid(test2))
plot(predict(test2), resid(test2))
```

Here, data is normalish but heteroskedastic. Log transform flips the problem. Conduct Welch's t-tests on untransformed data.

```{r}
# Cropped water vs C-amended
changeabund.trt.ph.welch1 <-  t.test(change_abund_corr ~ Amendment, data=growth.trt[growth.trt$Soil=="C3",], var.equal=FALSE)

# Successional water vs C-amended
changeabund.trt.ph.welch2 <-  t.test(change_abund_corr ~ Amendment, data=growth.trt[growth.trt$Soil=="S17",], var.equal=FALSE)

# Adjust p-values
changeabund.trt.ph.padj <- p.adjust(p <- c(changeabund.trt.ph.welch1$p.value, changeabund.trt.ph.welch2$p.value),
                    method = "BH", n=length(p))

# Show results
data.frame("Contrast" = c("Cropped, water vs C-amended", "Successional, water vs C-amended"),
           "p_value" = changeabund.trt.ph.padj)
```

Not enough power.

# Predicted 16S copy number

Typically, average 16S copy nunmber decreases with succession. This measurement is taken from whole communities, not just actively, growing members as in my study.

Hypothesis: Decrease in average 16S copy number among actively growing taxa in successional and water treatments compared to cropped and resource-amended treatments.

```{r}
# Visualize
growth.trt %>%
  ggplot(aes(x=Soil, y=n16S)) +
  geom_jitter(aes(shape=Amendment), width=0.1) +
  scale_shape_manual(values=c(19,1)) +
  theme_test()
```

**Statistics**

ANOVA

```{r}
growth.16s.lm = lm(n16S ~ Soil + Amendment + Soil:Amendment, data=growth.trt)
hist(resid(growth.16s.lm)) # check normality
plot(predict(growth.16s.lm), resid(growth.16s.lm)) # check heteroskedasticity
anova(growth.16s.lm)
```

# Create figure

```{r}
plot <- growth.trt %>%
  pivot_longer(c(start_day, change_abund_corr, k, n16S), names_to = "growth_metric") %>%
  mutate(Soil = fct_recode(Soil, Cropped = "C3", Successional = "S17"),
         Amendment = fct_recode(Amendment, `Water + Resources` = "Y", `Water only`= "N"),
         growth_metric = fct_relevel(growth_metric, c("start_day", "change_abund", "k", "n16S"))) %>%
  ggplot(aes(y=value, x=Soil)) +
  geom_jitter(aes(shape=Amendment), width=0.1) +
  scale_shape_manual(values=c(19,1)) +
  facet_wrap(~growth_metric, scales="free", nrow=1) +
  theme_classic() +
  theme(
        axis.title = element_blank(),
        text = element_text(size=9),
        strip.text.x = element_text(size = 10, hjust = 0),
        legend.position = "bottom",
        legend.title=element_blank(),
        strip.background = element_blank())
plot

ggsave(plot, filename="../figures/fig_growthparams_overall.svg", units="mm", width=180, height=60, device="svg")
```
