==> scripts/09_dunlopgr_groups.Rmd <==
---
title: "Dunlop growth rates - groups"
author: "Cassandra Wattenburger"
date: "3/8/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(results = "show")
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
```

**Goal of this script:** Growth rate distributions seem bimodal. See if groups can be explicitly modeled and explore relationships with other variables.

# Import libraries

```{r}
library(tidyverse)
library(mclust)

sessionInfo()

rm(list=ls())

set.seed(2) # mclust uses EM algorhythm which is iterative and random
```

# Import data

```{r}
growth <- readRDS("../rdata.files/gr_gr.paprica.clean.rds")
```

Average across replicates to get a single estimated growth rate for each taxa in each treatment.

```{r}
# Average across replicates for each ASV
growth.asv <- growth %>%
  group_by(Soil, Amendment, ASV) %>%
  summarize(k = mean(k),
            start_day = mean(start_day),
            end_day = mean(end_day),
            start_abund = mean(start_abund),
            end_abund = mean(end_abund)) %>% 
  ungroup()

# Including PAPRICA results
# Have to remove archaea because they weren't predicted by program (NAs)
growth.paprica.asv <- growth %>%
  na.omit() %>%
  group_by(Soil, Amendment, ASV) %>%
  summarize(k = mean(k),
            start_day = mean(start_day),
            end_day = mean(end_day),
            start_abund = mean(start_abund),
            end_abund = mean(end_abund),
            n16S = mean(n16S),
            genome_size = mean(genome_size)) %>% 
  ungroup()
```


Visualize distributions:

```{r}
# Averaged across replicates
growth.asv %>%
  ggplot(aes(x=log(k), color=Soil)) +
  geom_density() +
  facet_wrap(~Amendment) +
  theme_test()
```

# Fit models and cluster

### Successional water control

```{r}
# Subset
S17n <- growth.paprica.asv %>% 
  filter(Soil=="S17" & Amendment=="N")

# Fit univariate gaussian mixture model
S17n.mclust <- densityMclust(log(S17n$k))
summary(S17n.mclust)
plot(S17n.mclust$BIC)
S17n.mclust$parameters
plot(S17n.mclust, what = "density", data = log(S17n$k))

# Bootstrap CIs of parameters
S17n.boot <- MclustBootstrap(S17n.mclust, nboot=999, type="bs", verbose=FALSE)
summary(S17n.boot, what = "ci")
plot(S17n.boot, what = "mean")

# Add classifications
S17n.class <- S17n %>%
  add_column(classification = S17n.mclust$classification) %>%
  mutate(classification = as_factor(classification))

S17n.class %>% 
  ggplot(aes(x=log(k), color=classification)) +
  geom_histogram() +
  theme_test()
```


### Cropped water control

```{r}
# Subset
C3n <- growth.paprica.asv %>% 
  filter(Soil=="C3" & Amendment=="N")

# Fit univariate gaussian mixture model
C3n.mclust <- densityMclust(log(C3n$k))
summary(C3n.mclust)
plot(C3n.mclust$BIC)
C3n.mclust$parameters
plot(C3n.mclust, what = "density", data = log(C3n$k))

# Bootstrap CIs of parameters
C3n.boot <- MclustBootstrap(C3n.mclust, nboot=999, type="bs", verbose=FALSE)
summary(C3n.boot, what = "ci")
plot(C3n.boot, what = "mean")

# Add classifications
C3n.class <- C3n %>%
  add_column(classification = C3n.mclust$classification) %>%
  mutate(classification = as_factor(classification))

C3n.class %>% 
  ggplot(aes(x=log(k), color=classification)) +
  geom_histogram() +
  theme_test()
```


### Successional C-amended

```{r}
# Subset
S17y <- growth.paprica.asv %>% 
  filter(Soil=="S17" & Amendment=="Y")

# Fit univariate gaussian mixture model
S17y.mclust <- densityMclust(log(S17y$k))
summary(S17y.mclust)
plot(S17y.mclust$BIC)
S17y.mclust$parameters
plot(S17y.mclust, what = "density", data = log(S17y$k))

# Bootstrap CIs of parameters
S17y.boot <- MclustBootstrap(S17y.mclust, nboot=999, type="bs", verbose=FALSE)
summary(S17y.boot, what = "ci")
plot(S17y.boot, what = "mean")

# Add classifications
S17y.class <- S17y %>%
  add_column(classification = S17y.mclust$classification) %>%
  mutate(classification = as_factor(classification))

S17y.class %>% 
  ggplot(aes(x=log(k), color=classification)) +
  geom_histogram() +
  theme_test()
```

### Cropped C-amended

```{r}
# Subset
C3y <- growth.paprica.asv %>% 
  filter(Soil=="C3" & Amendment=="Y")

# Fit univariate gaussian mixture model
C3y.mclust <- densityMclust(log(C3y$k))
summary(C3y.mclust)
plot(C3y.mclust$BIC)
C3y.mclust$parameters
plot(C3y.mclust, what = "density", data = log(C3y$k))
C3y.boot <- MclustBootstrap(C3y.mclust, nboot=999, type="bs", verbose=FALSE)

# Add classifications
C3y.class <- C3y %>%
  add_column(classification = C3y.mclust$classification) %>%
  mutate(classification = as_factor(classification))
```

It looks like a model with a single distribution best fits this treatment, rather than 2 distributions.

# Cluster traits

```{r}
S17n_clusters <- S17n %>% 
  add_column(cluster=S17n.mclust$classification) %>% 
  mutate(cluster = if_else(cluster==1, "slow", "fast"))

S17y_clusters <- S17y %>% 
  add_column(cluster=S17y.mclust$classification) %>% 
  mutate(cluster = if_else(cluster==1, "slow", "fast"))

C3n_clusters <- C3n %>% 
  add_column(cluster=C3n.mclust$classification) %>% 
  mutate(cluster = if_else(cluster==1, "slow", "fast"))

C3y_clusters <- C3y %>% 
  add_column(cluster=C3y.mclust$classification) %>% 
  mutate(cluster = "fast")

# Combine
growth_clusters <- bind_rows(S17n_clusters, S17y_clusters) %>% 
  bind_rows(C3n_clusters) %>% 
  bind_rows(C3y_clusters)
```

```{r, eval=FALSE}
saveRDS(growth_clusters, "../rdata.files/gr_growth_clusters.rds")
```

## Specific growth rate

Visualize:

```{r}
# Soil
growth_clusters %>% 
  ggplot(aes(x=cluster, y=log(k))) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  facet_wrap(~Soil) +
  theme_test()

# Amendment
growth_clusters %>% 
  ggplot(aes(x=cluster, y=log(k))) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  facet_wrap(~Amendment) +
  theme_test()

# All
growth_clusters %>% 
  ggplot(aes(x=cluster, y=log(k), fill=Amendment)) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  facet_wrap(~Soil) +
  theme_test()
```

Statistics:

1. slow in successional resources vs water
2. fast in successional resources vs water
4. fast vs fast in cropped
5. slow in water amendment, cropped vs successional
6. fast in water amendment, cropped vs successional

Will do multiple test correction once appropriate tests are applied to each.

Remove intermediate cluster because not sure how comparable it is to fast or slow clusters.

```{r}
# Prep data for stats
growth_clusters_stat <- growth_clusters %>% 
  mutate(start_day = start_day+0.0001, # add small amount to all times to allow log
         change_abund_corr = (end_abund-start_abund)/n16S) %>% # change in abundance copy number corrected
  filter(cluster %in% c("fast", "slow")) # remove intermediate cluster from comparisons
```

```{r}
# Effect of amendment on clusters

# Slow successional water vs resources
lm1 <- lm(log(k) ~ Amendment, filter(growth_clusters, cluster=="slow" & Soil=="S17"))
hist(resid(lm1)) # normal 
plot(predict(lm1), resid(lm1)) # variance OK w/log transform
pval1 <- anova(lm1)$`Pr(>F)`[1]

# Fast successional water vs resources
lm2 <- lm(log(k) ~ Amendment, filter(growth_clusters, cluster=="fast" & Soil=="S17"))
hist(resid(lm2)) # normal 
plot(predict(lm2), resid(lm2)) # variance OK w/log transform
pval2 <- anova(lm2)$`Pr(>F)`[1]

# Fast vs intermediate cropped
# Intermediate has much more overlap with fast cluster ASVs in water only trt
lm4 <- lm(k ~ Amendment, filter(growth_clusters, cluster=="fast" & Soil=="C3"))
hist(resid(lm4)) # very right tailed w/o log
plot(predict(lm4), resid(lm4)) # variance OK w/o log
# I'll take non-parametric data over unequal variances
# Switch to kruskal wallis (non-parametric)
kw4 <- kruskal.test(k ~ Amendment, filter(growth_clusters, (cluster=="fast" & Soil=="C3")))
pval4 <- kw4$p.value

# Effect of soil history on clusters?

# Cropped vs successional, slow water
lm5 <- lm(log(k) ~ Soil, filter(growth_clusters, Amendment=="N" & cluster=="slow"))
hist(resid(lm5)) # normal w/ or w/o log 
plot(predict(lm5), resid(lm5)) # variance bad w/or w/o log
# Switch to Welch's t-test (unequal variances)
wt5 <- t.test(k ~ Soil, filter(growth_clusters, Amendment=="N" & cluster=="slow"))
pval5 <- wt5$p.value

# Cropped vs successional, fast water
lm6 <- lm(log(k) ~ Soil, filter(growth_clusters, Amendment=="N" & cluster=="fast"))
hist(resid(lm6)) # normal w/log
plot(predict(lm6), resid(lm6)) # variance bad w/ or w/o log
# switch to welch's t-test (unequal variances)
wt6 <- t.test(log(k) ~ Soil, filter(growth_clusters, Amendment=="N" & cluster=="fast"))
pval6 <- wt6$p.value

# Multiple test correction
pvals_df <- data.frame(test=c(1, 2, 3, 4, 5), pval=c(pval1, pval2, pval4, pval5, pval6))
pvals_adj <- p.adjust(pvals_df$pval, method="holm", n=5)
pvals_df <- bind_cols(pvals_df, pvals_adj=pvals_adj)
pvals_df %>% 
  mutate(sig = if_else(pvals_adj < 0.05, "*", "n.s."))
```

## Others

* Lag time
* Change in abundance
* n16S

Visualize:

```{r}
# Lag time
growth_clusters %>% 
  mutate(start_day = start_day+0.0001) %>% # add small amount to all times to allow log
  ggplot(aes(x=cluster, y=log(start_day), color=Amendment)) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  facet_wrap(~Soil) +
  theme_test()

# Change in abundance
growth_clusters %>% 
  mutate(change_abund_corr = (end_abund-start_abund)/n16S) %>% # change in abundance copy number corrected
  ggplot(aes(x=cluster, y=log(change_abund_corr), color=Amendment)) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  facet_wrap(~Soil) +
  theme_test()

# 16S copy number
growth_clusters %>% 
  ggplot(aes(x=cluster, y=log(n16S), color=Amendment)) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  facet_wrap(~Soil) +
  theme_test()
```

```{r}
# Lag time
lag_lm <- lm(log(start_day) ~cluster, growth_clusters_stat)
hist(resid(lag_lm)) # not normal w/ or w/o log
plot(predict(lag_lm), resid(lag_lm)) # variance is ok with log

# Use Kruskal Wallis
kruskal.test(log(start_day) ~ cluster, growth_clusters_stat)

# Change in abundance
chabund_lm <- lm(log(change_abund_corr) ~ cluster, growth_clusters_stat)
hist(resid(chabund_lm)) # normal with log transform
plot(predict(chabund_lm), resid(chabund_lm)) # variance OK
anova(chabund_lm)

# n16S
n16S_lm <- lm(log(n16S) ~ cluster, growth_clusters_stat)
hist(resid(n16S_lm)) # not normal w/ or w/o log transform
plot(predict(n16S_lm), resid(n16S_lm)) # variance good w/log transform

# Use Kruskal Wallis test
kruskal.test(log(n16S) ~ cluster, growth_clusters_stat)
```

Visualize stats comparisons:

```{r}
# Lag time
growth_clusters %>% 
  filter(cluster %in% c("fast", "slow")) %>% 
  mutate(start_day = start_day+0.0001) %>% # add small amount to all times to allow log
  ggplot(aes(x=cluster, y=log(start_day))) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  #facet_wrap(~Soil) +
  theme_test()

# Change in abundance
growth_clusters %>% 
  filter(cluster %in% c("fast", "slow")) %>% 
  mutate(change_abund_corr = (end_abund-start_abund)/n16S) %>% # change in abundance copy number corrected
  ggplot(aes(x=cluster, y=log(change_abund_corr))) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  #facet_wrap(~Soil) +
  theme_test()

# 16S copy number
growth_clusters %>% 
  filter(cluster %in% c("fast", "slow")) %>% 
  ggplot(aes(x=cluster, y=log(n16S))) +
  geom_boxplot() +
  geom_jitter(alpha=0.2) +
  #facet_wrap(~Soil) +
  theme_test()
```

# Ovelap between amendments

```{r}
# Identify overlapping ASVs in growth estimates between treatments across amendment levels

# Successional
S17y <- filter(growth_clusters, Soil=="S17", Amendment=="Y") %>% 
  select(ASV)
S17n <- filter(growth_clusters, Soil=="S17", Amendment=="N") %>% 
  select(ASV)
S17_overlap_asvs <- inner_join(S17y, S17n)
S17_overlap_asvs <- as.character(S17_overlap_asvs$ASV)
length(S17_overlap_asvs)

S17_growth_overlap <- growth_clusters %>% 
  filter(Soil=="S17" & ASV %in% S17_overlap_asvs) 

# Cropped
C3y <- filter(growth_clusters, Soil=="C3", Amendment=="Y") %>% 
  select(ASV)
C3n <- filter(growth_clusters, Soil=="C3", Amendment=="N") %>% 
  select(ASV)
C3_overlap_asvs <- inner_join(C3y, C3n)
C3_overlap_asvs <- as.character(C3_overlap_asvs$ASV)
length(C3_overlap_asvs)

C3_growth_overlap <- growth_clusters %>% 
  filter(Soil=="C3" & ASV %in% C3_overlap_asvs) 
```

Did k increase with amendment among overlapping taxa?

```{r}
S17_growth_overlap %>% 
  mutate(Amendment = as_factor(Amendment),
         Amendment = fct_relevel(Amendment, "N", "Y")) %>% 
  ggplot(aes(x=Amendment, y=k, color=cluster)) +
  geom_point() +
  geom_line(aes(group=ASV)) +
  labs(title="Successional") +
  theme_test()

C3_growth_overlap %>% 
  mutate(Amendment = as_factor(Amendment),
         Amendment = fct_relevel(Amendment, "N", "Y")) %>% 
  ggplot(aes(x=Amendment, y=k, color=cluster)) +
  geom_point() +
  geom_line(aes(group=ASV)) +
  labs(title="Cropped") +
  theme_test()
  
```

Statistics:

Paired t-tests

```{r}
# Successional
t.test(k ~ Amendment, S17_growth_overlap, paired=TRUE)

# Cropped
t.test(k ~ Amendment, C3_growth_overlap, paired=TRUE)
```

## Intermediate cluster

```{r}
inter_cluster <- growth_clusters %>% 
  filter(Soil=="C3")


  
```

# Figures

## Cluster densities

Code to create and plot clusters borrowed from https://github.com/jlw-ecoevo/eggo/blob/master/Scripts/RefSeq_Analysis.R

```{r}
# Generate clusters for graphing, scaled to proportion
C3n.clusters <- data.frame(x = seq(-5,0.5,0.01),
                    cl1 = dnorm(seq(-5,0.5,0.01),
                              mean = C3n.mclust$parameters$mean[1],
                              sd = sqrt(C3n.mclust$p$variance$sigmasq))*C3n.mclust$parameters$pro[1],
                    cl2 = dnorm(seq(-5,0.5,0.01),
                              mean = C3n.mclust$parameters$mean[2],
                              sd = sqrt(C3n.mclust$p$variance$sigmasq))*C3n.mclust$parameters$pro[2])

# Plot
plot1 <- growth.asv %>%
  filter(Soil=="C3", Amendment=="N") %>%
  ggplot() +
  geom_polygon(data=C3n.clusters, aes(x=x, y=cl1), linetype=2, alpha=0.6, color="black", fill="blue") +
  geom_polygon(data=C3n.clusters, aes(x=x, y=cl2), linetype=2, alpha=0.6, color="black", fill="red") +
  geom_density(aes(x=log(k))) +
  labs(title="Cropped, water", y="Kernel density", x="Specific growth rate (units), ln") +
  ylim(0,1) +
  theme_classic() +
  theme(plot.title = element_text(size=9),
        text = element_text(size=9),
        strip.background = element_blank(),
        axis.line = element_line(colour = "black"),
        axis.title = element_blank())
plot1
```

```{r}
# Generate clusters for graphing, scaled to proportion
C3y.clusters <- data.frame(x = seq(-5,0.5,0.01),
                    cl1 = dnorm(seq(-5,0.5,0.01),
                              mean = C3y.mclust$parameters$mean[1],
                              sd = sqrt(C3y.mclust$parameters$variance$sigmasq)))

# Plot
plot2 <- growth.asv %>%
  filter(Soil=="C3", Amendment=="Y") %>%
  ggplot() +
  geom_polygon(data=C3y.clusters, aes(x=x, y=cl1), linetype=2, alpha=0.6, color="black", fill="red") +
  geom_density(aes(x=log(k))) +
  labs(title="Cropped, resources", y="Kernel density", x="Specific growth rate (units), ln") +
  ylim(0,1) +
  theme_classic() +
  theme(plot.title = element_text(size=9),
        text = element_text(size=9),
        strip.background = element_blank(),
        axis.line = element_line(colour = "black"),
        axis.title = element_blank())
plot2
```

```{r}
# Generate clusters for graphing, scaled to proportion
S17n.clusters <- data.frame(x = seq(-5,0.5,0.01),
                    cl1 = dnorm(seq(-5,0.5,0.01),
                              mean = S17n.mclust$parameters$mean[1],
                              sd = sqrt(S17n.mclust$p$variance$sigmasq[1]))*S17n.mclust$parameters$pro[1],
                    cl2 = dnorm(seq(-5,0.5,0.01),
                              mean = S17n.mclust$parameters$mean[2],
                              sd = sqrt(S17n.mclust$p$variance$sigmasq[2]))*S17n.mclust$parameters$pro[2])

# Plot
plot3 <- growth.asv %>%
  filter(Soil=="S17", Amendment=="N") %>%
  ggplot() +
  geom_polygon(data=S17n.clusters, aes(x=x, y=cl1), linetype=2, alpha=0.6, color="black", fill="blue") +
  geom_polygon(data=S17n.clusters, aes(x=x, y=cl2), linetype=2, alpha=0.6, color="black", fill="red") +
  geom_density(aes(x=log(k))) +
  labs(title="Successional, water", y="Kernel density", x="Specific growth rate (units), ln") +
  ylim(0,0.7) +
  theme_classic() +
  theme(plot.title = element_text(size=9),
        text = element_text(size=9),
        strip.background = element_blank(),
        axis.line = element_line(colour = "black"),
        axis.title = element_blank())
plot3
```

```{r}
# Generate clusters for graphing, scaled to proportion
S17y.clusters <- data.frame(x = seq(-5,0.5,0.01),
                    cl1 = dnorm(seq(-5,0.5,0.01),
                              mean = S17y.mclust$parameters$mean[1],
                              sd = sqrt(S17y.mclust$p$variance$sigmasq[1]))*S17y.mclust$parameters$pro[1],
                    cl2 = dnorm(seq(-5,0.5,0.01),
                              mean = S17y.mclust$parameters$mean[2],
                              sd = sqrt(S17y.mclust$p$variance$sigmasq[2]))*S17y.mclust$parameters$pro[2])

# Plot
plot4 <- growth.asv %>%
  filter(Soil=="S17", Amendment=="Y") %>%
  ggplot() +
  geom_polygon(data=S17y.clusters, aes(x=x, y=cl1), linetype=2, alpha=0.6, color="black", fill="blue") +
  geom_polygon(data=S17y.clusters, aes(x=x, y=cl2), linetype=2, alpha=0.6, color="black", fill="red") +
  geom_density(aes(x=log(k))) +
  labs(title="Successional, resources", y="Kernel density", x="Specific growth rate (units), ln") +
  ylim(0,0.7) +
  theme_classic() +
  theme(plot.title = element_text(size=9),
        text = element_text(size=9),
        strip.background = element_blank(),
        axis.line = element_line(colour = "black"),
        axis.title = element_blank())
plot4
```

```{r}
library(cowplot)
figure <- plot_grid(plot1, plot2, plot3, plot4)
figure
```

```{r, eval=FALSE}
ggsave(figure, filename="../figures/fig_groups.svg", units="mm", width=85, height=85, device="svg")
```

## Slow and fast shift in successional

```{r}
plot_clusterk <- growth_clusters %>%
  mutate(Soil = if_else(Soil=="S17", "Successional", "Cropped"),
         Amendment = if_else(Amendment=="Y", "Resources + water", "Water only")) %>% 
  ggplot(aes(x=cluster, y=log(k), fill=Amendment)) +
  geom_boxplot() +
  facet_wrap(~Soil) +
  scale_fill_manual(values = c("#676366", "#FFFFFF")) +
  theme_test() +
  theme(axis.title = element_blank(),
        axis.text = element_text(size=10),
        text = element_text(size=9),
        strip.text.x = element_text(size = 11, hjust = 0),
        legend.title = element_blank(),
        legend.position = "right",
        legend.justification = "top",
        legend.text = element_text(size=10),
        strip.background = element_blank())
plot_clusterk
```

```{r, eval=FALSE}
ggsave(plot_clusterk, filename = "../figures/fig_clusterk.svg", units="mm", width=140, height=100, device="svg")
```

## BIC values

```{r}
C3n.bic <- data.frame(Clusters = 1:9, E = C3n.mclust$BIC[,1], V = C3n.mclust$BIC[,2]) %>% 
  pivot_longer(cols = -Clusters, names_to = "Model", values_to="BIC") %>% 
  add_column(Treatment = "Cropped, water")

C3y.bic <- data.frame(Clusters = 1:9, E = C3y.mclust$BIC[,1], V = C3y.mclust$BIC[,2]) %>% 
  pivot_longer(cols = -Clusters, names_to = "Model", values_to="BIC") %>% 
  add_column(Treatment = "Cropped, resources")

S17n.bic <- data.frame(Clusters = 1:9, E = S17n.mclust$BIC[,1], V = S17n.mclust$BIC[,2]) %>% 
  pivot_longer(cols = -Clusters, names_to = "Model", values_to="BIC") %>% 
  add_column(Treatment = "Successional, water")

S17y.bic <- data.frame(Clusters = 1:9, E = S17y.mclust$BIC[,1], V = S17y.mclust$BIC[,2]) %>% 
  pivot_longer(cols = -Clusters, names_to = "Model", values_to="BIC") %>% 
  add_column(Treatment = "Successional, resources")

all.bic <- bind_rows(C3n.bic, C3y.bic, S17n.bic, S17y.bic)

plot_bic <- all.bic %>% 
  ggplot(aes(shape=Model, x=Clusters, y=BIC)) +
  geom_point() +
  geom_line(aes(linetype=Model)) +
  facet_wrap(~Treatment, scales="free") +
  labs(x="Number of clusters") +
  scale_x_continuous(breaks = seq(1, 9, by = 1)) +
  theme_test()
plot_bic
```

```{r, eval=FALSE}
ggsave(plot_bic, filename = "../figures/suppfig_bic.svg", units="mm", width=140, height=100, device="svg")
```

## Bootstrapped cluster means

```{r}
# Succesional water only
S17n.boot.mean <- S17n.boot$mean %>% as.data.frame()
colnames(S17n.boot.mean) <- c("slow", "fast")
S17n.boot.mean <- S17n.boot.mean %>% 
  add_column(Soil="Successional",
             Amendment="Water only",
             intermediate=NA) %>% 
  select(Soil, Amendment, slow, fast, intermediate)

# Succesional resources + water
S17y.boot.mean <- S17y.boot$mean %>% as.data.frame()
colnames(S17y.boot.mean) <- c("slow", "fast")
S17y.boot.mean <- S17y.boot.mean %>% 
  add_column(Soil="Successional",
             Amendment="Resources + water",
             intermediate=NA) %>% 
  select(Soil, Amendment, slow, fast, intermediate)

# Cropped water only
C3n.boot.mean <- C3n.boot$mean %>% as.data.frame()
colnames(C3n.boot.mean) <- c("slow", "fast")
C3n.boot.mean <- C3n.boot.mean %>% 
  add_column(Soil="Cropped",
             Amendment="Water only",
             intermediate=NA) %>% 
  select(Soil, Amendment, slow, fast, intermediate)

# Cropped resources + water
C3y.boot.mean <- C3y.boot$mean %>% as.data.frame()
colnames(C3y.boot.mean) <- c("intermediate")
C3y.boot.mean <- C3y.boot.mean %>% 
  add_column(Soil="Cropped",
             Amendment="Resources + water",
             fast=NA,
             slow=NA) %>% 
  select(Soil, Amendment, slow, fast, intermediate)

# Combine
boot_mean_all <- add_row(S17n.boot.mean, S17y.boot.mean) %>% 
  add_row(C3n.boot.mean) %>% 
  add_row(C3y.boot.mean)

# Long format
boot_mean_all <- boot_mean_all %>% 
  pivot_longer(slow:intermediate, names_to="cluster", values_to="mean")
```

Visualize:

```{r}
plot_bootstrap <- boot_mean_all %>% 
  mutate(name=paste0(Soil, ", ", Amendment, ", ", cluster)) %>% 
  ggplot(aes(x=Soil, y=mean, color=cluster)) +
  geom_boxplot() +
  facet_wrap(~Amendment) +
  labs(x="", y="Bootstrapped means (ln specific growth rate) ") +
  scale_color_manual(values=c("blue", "purple", "red")) +
  theme_test() 
plot_bootstrap
```

```{r, eval=FALSE}
ggsave(plot_bootstrap, filename = "../figures/suppfig_meanboot.svg", units="mm", width=140, height=100, device="svg")
```
